Issues 356 - Refactoring bash command
