+++ mycroft/client/wifisetup/FileUtils.py
@@ -59,32 +59,6 @@ class WriteFileTemplate(threading.Thread):
             return 1
 
 
-def write_wpa_supplicant_conf(ssid, passphrase):
-    template = """country={country}
-ctrl_interface=/var/run/wpa_supplicant
-update_config=1
-network={b1}
-ssid={ssid}
-psk={passphrase}
-key_mgmt=WPA-PSK
-{b2}
-"""
-    context = {
-        "b1": '{',
-        "b2": '}',
-        "country": 'US',
-        "ssid": ssid,
-        "passphrase": passphrase
-    }
-    wpa_conf_write = WriteFileTemplate(
-        'Write Wireless Settings to wpa_supplicant.conf ',
-        '/etc/wpa_supplicant/wpa_supplicant.conf',
-        template,
-        context)
-    wpa_conf_write.start()
-    wpa_conf_write.join()
-
-
 def write_hostapd_conf(interface, driver, ssid, channel):
     template = """interface={interface}
 driver={driver}
@@ -147,43 +121,6 @@ DAEMON_CONF="{config_file_path}"
     default_hostapd_write.join()
 
 
-def write_network_interfaces(wdev, vdev, vdev_address, vdev_hwaddress):
-    template = """source-directory /etc/network/interfaces.d
-
-auto lo
-iface lo inet loopback
-
-auto eth0
-allow-hotplug eth0
-iface eth0 inet dhcp
-
-auto {wdev}
-allow-hotplug {wdev}
-iface {wdev} inet dhcp
-    wpa-conf /etc/wpa_supplicant/wpa_supplicant.conf
-
-auto {vdev}
-allow-hotplug {vdev}
-iface {vdev} inet static
-    address {vdev_address}
-    netmask 255.255.255.0
-    hwaddress ether {vdev_hwaddress}
-"""
-    context = {
-        "wdev": wdev,
-        "vdev": vdev,
-        "vdev_address": vdev_address,
-        "vdev_hwaddress": vdev_hwaddress
-    }
-    network_interfaces_write = WriteFileTemplate(
-        'Write Network interface settings to ',
-        '/etc/network/interfaces',
-        template,
-        context)
-    network_interfaces_write.start()
-    network_interfaces_write.join()
-
-
 def write_dnsmasq(interface, server, dhcp_range_start, dhcp_range_end):
     template = """interface={interface}
 bind-interfaces
@@ -205,41 +142,6 @@ address=/#/172.24.1.1
                                        template, context)
     dns_conf_write.start()
     dns_conf_write.join()
-
-
-def backup_system_files():
-    backup_path = '/tmp'
-    etc_network_interfaces = CopyFile('Backup - Network/Interfaces: ',
-                                      '/etc/network/', 'interfaces',
-                                      backup_path, 'out.net')
-    etc_network_interfaces.start()
-    etc_network_interfaces.join()
-
-    etc_wpa_supplicant = CopyFile('Backup - Network/Interfaces: ',
-                                  '/etc/wpa_supplicant/',
-                                  'wpa_supplicant.conf',
-                                  backup_path, 'out.wpa')
-    etc_wpa_supplicant.start()
-    etc_wpa_supplicant.join()
-
-    etc_hostapd = CopyFile('Backup - HostAPD: ', '/etc/hostapd/',
-                           'hostapd.conf', backup_path, 'out.ap')
-    etc_hostapd.start()
-    etc_hostapd.join()
-
-    etc_default_hostapd = CopyFile('Backup - Default/HostAPD: ',
-                                   '/etc/default/', 'hostapd',
-                                   backup_path, 'out.default.ap')
-    etc_default_hostapd.start()
-    etc_default_hostapd.join()
-
-    etc_dnsmasq = CopyFile('Backup - DNSMasq: ',
-                           '/etc/',
-                           'dnsmasq.conf',
-                           backup_path,
-                           'out.dnsmasq')
-    etc_dnsmasq.start()
-    etc_dnsmasq.join()
 
 
 def restore_system_files():
@@ -273,18 +175,3 @@ def restore_system_files():
         'dnsmasq.conf')
     etc_dnsmasq.start()
     etc_dnsmasq.join()
-
-
-def ap_mode_config():
-    backup_system_files()
-    write_network_interfaces(
-        'wlan0', 'uap0', '172.24.1.1', 'bc:5f:f4:be:7d:0a')
-    write_hostapd_conf('uap0', 'nl80211', 'Mycroft-bing-bong-boom', 6)
-
-
-def ap_mode_deconfig():
-    restore_system_files()
-
-
-def client_mode_config(iface, ssid, passphrase):
-    write_wpa_supplicant_conf(ssid, passphrase)
